product-service/src/test/java/com/company/productservice/controller/ProductControllerTest.java
package com.company.productservice.controller;

import com.company.productservice.DTO.Product;
import com.company.productservice.service.ServiceLayer;
import com.company.productservice.viewModels.ProductView;
import com.fasterxml.jackson.databind.ObjectMapper;
import org.junit.Before;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.autoconfigure.ImportAutoConfiguration;
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
import org.springframework.boot.test.mock.mockito.MockBean;
import org.springframework.cloud.autoconfigure.RefreshAutoConfiguration;
import org.springframework.http.MediaType;
import org.springframework.test.context.junit4.SpringRunner;
import org.springframework.test.web.servlet.MockMvc;

import java.math.BigDecimal;
import java.util.ArrayList;
import java.util.List;

import static org.junit.Assert.*;
import static org.mockito.Mockito.doReturn;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.delete;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;

@RunWith(SpringRunner.class)
@WebMvcTest(ProductController.class)
public class ProductControllerTest {


    @Autowired
    MockMvc mockMvc;

    @MockBean
    ServiceLayer service;

    ProductView productToCreate;
    ProductView fromDB;
    List<ProductView> productViewList;

    // Mapper to turn Java objects into JSON and vice versa
    private ObjectMapper mapper = new ObjectMapper();

    @Before
    public void setUp() throws Exception {
        // setup test objects
        productToCreate = new ProductView();
        productToCreate.setName("Product");
        productToCreate.setDescription("Desc");
        productToCreate.setListPrice(new BigDecimal("19.99").setScale(2));
        productToCreate.setUnitCost(new BigDecimal("4.99").setScale(2));

        //product from DB
        fromDB = new ProductView();
        fromDB.setId(1);
        fromDB.setName("Product");
        fromDB.setDescription("Desc");
        fromDB.setListPrice(new BigDecimal("19.99").setScale(2));
        fromDB.setUnitCost(new BigDecimal("4.99").setScale(2));

        productViewList = new ArrayList<>();
        productViewList.add(fromDB);

        // fire up
        setUpServiceMock();
    }

    @Test
    public void shouldAddProduct() throws Exception {
        //Assemble
        String inputJson = mapper.writeValueAsString(productToCreate);
        String outputJson = mapper.writeValueAsString(fromDB);
        System.out.println(inputJson);
        System.out.println(outputJson);
        //Act
        mockMvc.perform(post("/products")
                .contentType(MediaType.APPLICATION_JSON)
                .content(inputJson)
                .accept(MediaType.APPLICATION_JSON))
                //Assert
                .andExpect(status().isCreated())
                .andExpect(content().json(outputJson));
    }

    @Test
    public void shouldThrow422WhenEmptyFields() throws Exception {
        //Assemble
        ProductView invalid = new ProductView();

        String inputJson = mapper.writeValueAsString(invalid);


        //Act
        mockMvc.perform(post("/products")
                .contentType(MediaType.APPLICATION_JSON)
                .content(inputJson)
                .accept(MediaType.APPLICATION_JSON))
                //Assert
                .andExpect(status().isUnprocessableEntity());
    }

    @Test
    public void shouldGetAllProducts()  throws Exception {
        String outputJson = mapper.writeValueAsString(productViewList);

        //Act
        mockMvc.perform(get("/products"))
                //Assert
                .andExpect(status().isOk())
                .andExpect(content().json(outputJson));
    }

    @Test
    public void shouldGetProduct() throws Exception {
        String outputJson = mapper.writeValueAsString(fromDB);

        //Act
        mockMvc.perform(get("/products/1"))
                //Assert
                .andExpect(status().isOk())
                .andExpect(content().json(outputJson));
    }

    @Test
    public void shouldReturn404WhenNotFound() throws Exception {
        //Act
        mockMvc.perform(get("/products/100"))
                //Assert
                .andExpect(status().isNotFound());
    }

    @Test
    public void shouldUpdateProduct() throws Exception {
        //Assemble
        String inputJson = mapper.writeValueAsString(fromDB);
        //Act
        mockMvc.perform(put("/products/1")
                .contentType(MediaType.APPLICATION_JSON)
                .content(inputJson)
                .accept(MediaType.APPLICATION_JSON))
                //Assert
                .andExpect(status().isOk());
    }

    @Test
    public void shouldReturn422WhenUpdatingWithInvalidId() throws Exception {
        //Assemble
        String inputJson = mapper.writeValueAsString(fromDB);
        //Act
        mockMvc.perform(put("/products/100")
                .contentType(MediaType.APPLICATION_JSON)
                .content(inputJson)
                .accept(MediaType.APPLICATION_JSON))
                //Assert
                .andExpect(status().isUnprocessableEntity());
    }

    @Test
    public void shouldDeleteProduct() throws Exception {
        //Act
        mockMvc.perform(delete("/products/1"))
                //Assert
                .andExpect(status().isNoContent());
    }
// forgot to incorporate NotFoundException in deleteProduct method of controller.
// Will incorporate later.

//    @Test
//    public void shouldReturn404WhenDeletingNonExistingProduct() throws Exception {
//        //Act
//        mockMvc.perform(delete("/products/100"))
//                //Assert
//                .andExpect(status().isNotFound());
//    }

    // helper method
    private void setUpServiceMock() {
        doReturn(fromDB).when(service).newProduct(productToCreate);
        doReturn(fromDB).when(service).fetchProduct(1);
        doReturn(productViewList).when(service).fetchAllProducts();
    }
}